+++ src/routes/auth/index.ts
@@ -23,8 +23,8 @@ const router = Router()
 router.use(nocache())
 
 router.use((req, res, next) => {
-  if (!AUTHENTICATION.ENABLE) {
-    return res.boom.badImplementation(`Please set the AUTH_ENABLE env variable to true to use the auth routes.`)
+  if (!AUTHENTICATION.ENABLED) {
+    return res.boom.badImplementation(`Please set the AUTH_ENABLED env variable to true to use the auth routes.`)
   } else {
     return next();
   }