+++ src/shared/config/storage.ts
@@ -4,11 +4,11 @@ import { castBooleanEnv } from './utils'
  * * Storage Settings
  */
 export const STORAGE = {
-  get ENABLE() {
-    return castBooleanEnv('STORAGE_ENABLE', true)
+  get ENABLED() {
+    return castBooleanEnv('STORAGE_ENABLED') || castBooleanEnv('STORAGE_ENABLE') || true
   },
   get S3_SSL_ENABLED() {
-    return castBooleanEnv('S3_SSL_ENABLED', true)
+    return castBooleanEnv('S3_SSL_ENABLED') || castBooleanEnv('S3_SSL_ENABLE') || true
   },
   get S3_BUCKET() {
     return process.env.S3_BUCKET || ''